+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Plugins can now be reloaded
 - Adds button to maximize graph widgets
 - Adds button to pause all experiments
+- Allows setting the node and edge sizes in the graph widgets
 - New model plugin: population growth
 - New model plugin: Conway's Game of Life cellular automaton